# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"API","description":"API description","contact":{"email":"admin@admin.com"},"license":{"name":"Apache 2.0","url":"http:\/\/www.apache.org\/licenses\/LICENSE-2.0.html"},"version":"1.0.0"},"paths":{"\/api\/login":{"post":{"tags":["Login"],"summary":"Login User","description":"Login User","operationId":"a3b306d14572d1f4bd6c064b3233e7b8","requestBody":{"description":"Provide All Info Below","required":true,"content":{"application\/json":{"schema":{"required":["email","password"],"properties":{"email":{"type":"email","format":"text","example":"mercedes68@example.org"},"password":{"type":"string","format":"text","example":"123456"}},"type":"object"}}}},"responses":{"200":{"description":"Login Success","content":{"application\/json":{"schema":{"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Token"}},"type":"object"}}}},"500":{"description":"Log Information store failed","content":{"application\/json":{"schema":{"properties":{"status":{"type":"string","example":"error"},"message":{"type":"string","example":"Some issue found"}},"type":"object"}}}}}}},"\/api\/register":{"post":{"tags":["Register"],"summary":"Register User","description":"Register User","operationId":"8a56853624e025573120a09a4c75d468","requestBody":{"description":"Provide All Info Below","required":true,"content":{"application\/json":{"schema":{"required":["name","email","password"],"properties":{"name":{"type":"string","format":"text","example":"test123"},"email":{"type":"email","format":"text","example":"test@example.org"},"password":{"type":"string","format":"text","example":"123456"}},"type":"object"}}}},"responses":{"200":{"description":"Login Success","content":{"application\/json":{"schema":{"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Registration Successfull"}},"type":"object"}}}},"500":{"description":"Log Information store failed","content":{"application\/json":{"schema":{"properties":{"status":{"type":"string","example":"error"},"message":{"type":"string","example":"Some issue found"}},"type":"object"}}}}}}}},"components":{"securitySchemes":{"bearer":{"type":"apiKey","description":"Enter token in format (Bearer <token>)","name":"Authorization","in":"header","required":true}}}}